Adding GCP VPC Final
